refactoring: split search cache classes to several files
