+++ src/Microsoft.TemplateSearch.Common/PublicAPI.Unshipped.txt
@@ -26,5 +26,16 @@ Microsoft.TemplateSearch.Common.TemplateSearchCoordinator.SearchAsync(System.Fun
 Microsoft.TemplateSearch.Common.TemplateSearchCoordinator.TemplateSearchCoordinator(Microsoft.TemplateEngine.Abstractions.IEngineEnvironmentSettings! environmentSettings, System.Collections.Generic.IReadOnlyDictionary<string!, System.Func<object!, object!>!>? additionalDataReaders = null) -> void
 Microsoft.TemplateSearch.Common.TemplateSearchData
 Microsoft.TemplateSearch.Common.TemplateSearchData.AdditionalData.get -> System.Collections.Generic.IDictionary<string!, object!>!
+Microsoft.TemplateSearch.Common.TemplateSearchData.Author.get -> string?
+Microsoft.TemplateSearch.Common.TemplateSearchData.Classifications.get -> System.Collections.Generic.IReadOnlyList<string!>!
+Microsoft.TemplateSearch.Common.TemplateSearchData.Description.get -> string?
+Microsoft.TemplateSearch.Common.TemplateSearchData.GroupIdentity.get -> string?
+Microsoft.TemplateSearch.Common.TemplateSearchData.Identity.get -> string!
+Microsoft.TemplateSearch.Common.TemplateSearchData.Name.get -> string!
+Microsoft.TemplateSearch.Common.TemplateSearchData.Parameters.get -> System.Collections.Generic.IReadOnlyList<Microsoft.TemplateEngine.Abstractions.ITemplateParameter!>!
+Microsoft.TemplateSearch.Common.TemplateSearchData.Precedence.get -> int
+Microsoft.TemplateSearch.Common.TemplateSearchData.ShortNameList.get -> System.Collections.Generic.IReadOnlyList<string!>!
+Microsoft.TemplateSearch.Common.TemplateSearchData.TagsCollection.get -> System.Collections.Generic.IReadOnlyDictionary<string!, string!>!
 Microsoft.TemplateSearch.Common.TemplateSearchData.TemplateSearchData(Microsoft.TemplateEngine.Abstractions.ITemplateInfo! templateInfo, System.Collections.Generic.IDictionary<string!, object!>? data = null) -> void
+Microsoft.TemplateSearch.Common.TemplateSearchData.ThirdPartyNotices.get -> string?
 static Microsoft.TemplateSearch.Common.Components.AllComponents.get -> System.Collections.Generic.IReadOnlyList<(System.Type! Type, Microsoft.TemplateEngine.Abstractions.IIdentifiedComponent! Instance)>!
